+++ models/ItemColor.js
@@ -4,8 +4,6 @@ const { Model } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
   class ItemColor extends Model {
     static associate(models) {
-      this.hasMany(models.OptionImage, { foreignKey: 'itemColorId' });
-      this.hasMany(models.OptionSize, { foreignKey: 'itemColorId' });
       this.belongsTo(models.Item, { foreignKey: 'itemId' });
     }
   }
